# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/equippableItems": {
            "get": {
                "tags": [
                    "EquippableItems"
                ],
                "summary": "Get all equippable items",
                "description": "Returns a list of all equippable items",
                "operationId": "getAllEquippableItems",
                "responses": {
                    "200": {
                        "description": "A list of equippable items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "equippableItems list stringify"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "EquippableItems"
                ],
                "summary": "Create a random equippable item",
                "description": "Creates a random equippable item based on the given level and ownerId",
                "operationId": "createRandomItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "level",
                                    "ownerId"
                                ],
                                "properties": {
                                    "level": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "ownerId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "rarityId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "blueprintId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "ownerId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "randomFactor": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 0.5
                                        },
                                        "damage": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 10.5
                                        },
                                        "health": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 20.5
                                        },
                                        "getRarity": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Rare"
                                                },
                                                "hexColor": {
                                                    "type": "string",
                                                    "example": "#FF0000"
                                                },
                                                "multiplier": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1.5
                                                },
                                                "levelRequiredToDrop": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "getBlueprint": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Sword"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "A sharp sword"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "weapon"
                                                },
                                                "baseDamage": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "baseHealth": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "requiredLevel": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No available rarity or blueprint for the given level"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/equippableItems/{id}": {
            "get": {
                "tags": [
                    "EquippableItems"
                ],
                "summary": "Get equippable item by ID",
                "description": "Returns a single equippable item by its ID",
                "operationId": "getEquippableItemById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the equippable item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equippable item details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "equippableItem details stringify"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}